Address comments * Move drop report tests inside of existing `int_traffic_trace` linerate test
